requirements.txt
PyMySQL==1.0.2
six==1.16.0
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8

cffi==1.15.0
colorama==0.4.4
cryptography==35.0.0
distlib==0.3.3
dominate==2.6.0
filelock==3.3.2
Flask-Bcrypt==0.7.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-WTF==1.0.0
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
pipenv==2021.11.9
platformdirs==2.4.0
protobuf==3.19.1
pycparser==2.21
six==1.16.0
SQLAlchemy==1.4.26
virtualenv==20.10.0
virtualenv-clone==0.5.7
visitor==0.1.3
Werkzeug==2.0.2
WTForms==3.0.0
gunicorn==19.9.0
